move controller logic to service class
